Move model logic to separate class
